🐛구조 및 README.md 수정
